[#23] fix: db 구조 수정
- 모든 도메인의 id타입을 Long으로 변경
- User와 Card를 Many to one 관계로 수정
- User와 Activity를 Many to one 관계로 수정

+++ BE/src/main/java/com/codesquad/todo/domain/CardRepository.java
@@ -2,5 +2,5 @@ package com.codesquad.todo.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CardRepository extends CrudRepository<Card, Integer> {
+public interface CardRepository extends CrudRepository<Card, Long> {
 }
