+++ BE/src/main/java/com/codesquad/todo/domain/SectionRepository.java
@@ -2,5 +2,5 @@ package com.codesquad.todo.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface SectionRepository extends CrudRepository<Section, Integer> {
+public interface SectionRepository extends CrudRepository<Section, Long> {
 }
